+++ example/android/app/src/main/kotlin/com/example/flutter_permission_handler_plus_example/MainActivity.kt
@@ -1,0 +1,5 @@
+package com.example.flutter_permission_handler_plus_example
+
+import io.flutter.embedding.android.FlutterActivity
+
+class MainActivity : FlutterActivity()